# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /withdrawals/overview: {"get":{"operationId":"getWithdrawalOverview","tags":["withdrawals"],"summary":"Read existing wallet coins, reserved withdrawals and enrollment policy","description":"Creator-scoped authenticated wallet withdrawal; no automatic payments.","security":[{"sessionBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalOverview"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}}}}
  /withdrawals/recipient: {"get":{"operationId":"getWithdrawalRecipient","tags":["withdrawals"],"summary":"Read saved legal/contact details; banking details remain with Remitly","description":"Creator-scoped authenticated wallet withdrawal; no automatic payments.","security":[{"sessionBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/WithdrawalRecipient"}],"nullable":true}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}}},"put":{"operationId":"saveWithdrawalRecipient","tags":["withdrawals"],"summary":"Save creator recipient contact details","description":"Creator-scoped authenticated wallet withdrawal; no automatic payments.","security":[{"sessionBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalRecipient"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalRecipientInput"}}}}}}
  /withdrawals: {"post":{"operationId":"requestWithdrawal","tags":["withdrawals"],"summary":"Atomically reserve 6,000 existing wallet coins for USD 15 gross; await exact quote confirmation","description":"Creator-scoped authenticated wallet withdrawal; no automatic payments.","security":[{"sessionBearer":[]}],"responses":{"201":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatorWithdrawal"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalRequestInput"}}}}}}
  /withdrawals/{id}: {"get":{"operationId":"getWithdrawalDetail","tags":["withdrawals"],"summary":"Read creator-owned withdrawal status and timeline","description":"Creator-scoped authenticated wallet withdrawal; no automatic payments.","security":[{"sessionBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}]}}
  /withdrawals/{id}/approve-quote: {"post":{"operationId":"approveWithdrawalQuote","tags":["withdrawals"],"summary":"Creator independently approves current exact signed-in quote hash","description":"Creator-scoped authenticated wallet withdrawal; no automatic payments.","security":[{"sessionBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalQuoteApprovalInput"}}}}}}
  /withdrawals/{id}/cancel: {"post":{"operationId":"cancelUnpreparedWithdrawal","tags":["withdrawals"],"summary":"Cancel and refund wallet coins only before any provider attempt","description":"Creator-scoped authenticated wallet withdrawal; no automatic payments.","security":[{"sessionBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}]}}
  /withdrawals/{id}/statement: {"get":{"operationId":"downloadWithdrawalStatement","tags":["withdrawals"],"summary":"Download creator-owned truthful payout statement","description":"Creator-scoped authenticated wallet withdrawal; no automatic payments.","security":[{"sessionBearer":[]}],"responses":{"200":{"description":"Plain-text earnings statement; pending payments are labeled","headers":{"Content-Disposition":{"schema":{"type":"string"}}},"content":{"text/plain":{"schema":{"type":"string"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}]}}
  /admin-data/withdrawals: {"get":{"operationId":"listAdminWithdrawals","tags":["withdrawals"],"summary":"Read account-scoped payout queue, exceptions and latest attempt deadlines","description":"Enabled owner bearer and production MFA required. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminWithdrawalQueue"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}}}}
  /admin-data/withdrawals/enrollment-preview: {"get":{"operationId":"previewWithdrawalEnrollment","tags":["withdrawals"],"summary":"Preview selected creator wallet; bought, gifted and granted coins are all redeemable","description":"Enabled owner bearer and production MFA required. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalEnrollmentPreview"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","minimum":1}}]}}
  /admin-data/withdrawals/enroll: {"post":{"operationId":"enrollWithdrawalCreator","tags":["withdrawals"],"summary":"Enable one creator using matching wallet snapshot; never mint or backfill earnings","description":"Enabled owner bearer and production MFA required. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalEnrollmentResult"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalEnrollmentInput"}}}}}}
  /admin-data/withdrawals/pause: {"post":{"operationId":"setWithdrawalPreparationPause","tags":["withdrawals"],"summary":"Pause new preparation while reconciliation remains available","description":"Enabled owner bearer and production MFA required. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalPauseResult"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalPauseInput"}}}}}}
  /admin-data/withdrawals/{id}: {"get":{"operationId":"getAdminWithdrawalDetail","tags":["withdrawals"],"summary":"Read protected attempts, legal contacts, evidence and audit identities","description":"Enabled owner bearer and production MFA required. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminWithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}]}}
  /admin-data/withdrawals/{id}/quote: {"post":{"operationId":"recordWithdrawalQuote","tags":["withdrawals"],"summary":"Record actual signed-in provider quote and invalidate earlier approval/checks","description":"Enabled owner bearer and production MFA required. Additionally requires configured maker operator role. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalQuoteInput"}}}}}}
  /admin-data/withdrawals/{id}/prepare: {"post":{"operationId":"beginWithdrawalPreparation","tags":["withdrawals"],"summary":"Allocate durable attempt before possible provider creation; preparation pause applies","description":"Enabled owner bearer and production MFA required. Additionally requires configured maker operator role. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalPrepareResult"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalPrepareInput"}}}}}}
  /admin-data/withdrawals/{id}/preparation: {"post":{"operationId":"recordWithdrawalPreparation","tags":["withdrawals"],"summary":"Record first-time link plan or one-time draft with auto-send off","description":"Enabled owner bearer and production MFA required. Additionally requires configured maker operator role. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalPreparationInput"}}}}}}
  /admin-data/withdrawals/{id}/check: {"post":{"operationId":"checkWithdrawalPreparation","tags":["withdrawals"],"summary":"Independently check exact attempt version; maker cannot self-check","description":"Enabled owner bearer and production MFA required. Additionally requires configured checker operator role. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalCheckInput"}}}}}}
  /admin-data/withdrawals/{id}/release: {"post":{"operationId":"recordHumanWithdrawalRelease","tags":["withdrawals"],"summary":"Record human approval and actual external Remitly action; does not send payment","description":"Enabled owner bearer and production MFA required. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalReleaseInput"}}}}}}
  /admin-data/withdrawals/{id}/decline: {"post":{"operationId":"declineWithdrawal","tags":["withdrawals"],"summary":"Final human decline; refund before attempt or retain reserve pending provider cancellation","description":"Enabled owner bearer and production MFA required. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDeclineInput"}}}}}}
  /admin-data/withdrawals/{id}/unknown: {"post":{"operationId":"markWithdrawalInvestigation","tags":["withdrawals"],"summary":"Record unknown/expired outcome without releasing coins or allowing retries","description":"Enabled owner bearer and production MFA required. Additionally requires configured reconciler operator role. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalInvestigationInput"}}}}}}
  /admin-data/withdrawals/{id}/reconcile: {"post":{"operationId":"reconcileWithdrawal","tags":["withdrawals"],"summary":"Record verified provider observations; wallet refunds require authoritative return","description":"Enabled owner bearer and production MFA required. Additionally requires configured reconciler operator role. Final approve/decline and release decisions remain human-owned. No operation sends payments to Remitly.","security":[{"adminBearer":[]}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalDetail"}}}},"400":{"description":"Invalid request or provider evidence"},"401":{"description":"Explicit authenticated session bearer required"},"403":{"description":"Creator enrollment or enabled owner, production MFA and configured operator role required"},"404":{"description":"Withdrawal not found in authenticated creator/provider-account scope"},"409":{"description":"Insufficient wallet coins, unresolved attempt, stale quote/check, duplicate provider identity or unsafe state change"},"503":{"description":"Withdrawal service or provider account configuration unavailable"}},"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WithdrawalReconcileInput"}}}}}}
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
    adminBearer:
      type: http
      scheme: bearer
  schemas:
...
    WithdrawalRecipient: {"type":"object","required":["legalFirstName","legalLastName","countryCode","email","phone"],"properties":{"legalFirstName":{"type":"string"},"legalLastName":{"type":"string"},"secondSurname":{"type":"string","nullable":true},"countryCode":{"type":"string","pattern":"^[A-Z]{2}$"},"email":{"type":"string"},"phone":{"type":"string"},"revision":{"type":"integer"},"status":{"type":"string","enum":["contact_saved"]}}}
    WithdrawalRecipientInput: {"type":"object","required":["legalFirstName","legalLastName","countryCode","email","phone"],"properties":{"legalFirstName":{"type":"string","minLength":1,"maxLength":100},"legalLastName":{"type":"string","minLength":1,"maxLength":100},"secondSurname":{"type":"string","maxLength":100},"countryCode":{"type":"string","pattern":"^[A-Z]{2}$"},"email":{"type":"string","maxLength":254},"phone":{"type":"string","pattern":"^\\+\\d{8,15}$"}},"additionalProperties":false}
...
    WithdrawalDetail: {"allOf":[{"$ref":"#/components/schemas/CreatorWithdrawal"},{"type":"object","required":["history"],"properties":{"history":{"type":"array","items":{"$ref":"#/components/schemas/WithdrawalHistoryEntry"}}}}]}
    WithdrawalOverview: {"type":"object","required":["enrolled","balances","policy","recipient","withdrawals"],"properties":{"enrolled":{"type":"boolean"},"balances":{"$ref":"#/components/schemas/WithdrawalBalances"},"policy":{"$ref":"#/components/schemas/WithdrawalPolicy"},"recipient":{"allOf":[{"$ref":"#/components/schemas/WithdrawalRecipient"}],"nullable":true},"withdrawals":{"type":"array","items":{"$ref":"#/components/schemas/CreatorWithdrawal"}}}}
    WithdrawalRequestInput: {"type":"object","required":["methodId","withdrawalCents","idempotencyKey"],"properties":{"methodId":{"type":"string","minLength":1,"maxLength":250},"withdrawalCents":{"type":"integer","enum":[1500]},"idempotencyKey":{"type":"string","minLength":1,"maxLength":100}},"additionalProperties":false}
...